+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/IGeneratedEnumTests.java
@@ -1,7 +1,9 @@
 package io.github.chrimle.o2jrm.tests;
 
+import com.fasterxml.jackson.annotation.JsonValue;
 import io.github.chrimle.o2jrm.GeneratedSource;
 import io.github.chrimle.o2jrm.utils.CustomAssertions;
+import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -30,6 +32,46 @@ public abstract class IGeneratedEnumTests {
         @DisplayName("Generates an `enum` class")
         void whenIsEnumThenGeneratedClassIsEnumClass(final GeneratedSource generatedSource) {
           CustomAssertions.assertClassIsEnumClass(generatedSource.getClassUnderTest());
+        }
+
+        @Nested
+        @DisplayName("Testing `getValue()` method")
+        class IGetValueMethodTests {
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName("Generated `enum` classes always has `getValue()`-method")
+          void alwaysGenerateGetValueMethod(final GeneratedSource generatedSource) {
+            CustomAssertions.assertClassHasMethod(generatedSource.getClassUnderTest(), "getValue");
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "When `serializationLibrary` is `jackson` Then `getValue()`-method IS annotated with `@JsonValue`")
+          void whenSerializationLibraryIsJacksonThenGetValueMethodIsAnnotatedWithJsonValue(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.isSerializationLibraryJackson());
+
+            CustomAssertions.assertMethodIsAnnotatedWith(
+                CustomAssertions.assertClassHasMethod(
+                    generatedSource.getClassUnderTest(), "getValue"),
+                JsonValue.class);
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "When `serializationLibrary` is NOT `jackson` Then `getValue()`-method is NOT annotated with `@JsonValue`")
+          void whenSerializationLibraryIsNotJacksonThenGetValueMethodIsNotAnnotatedWithJsonValue(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.isSerializationLibraryJackson());
+
+            CustomAssertions.assertMethodIsNotAnnotatedWith(
+                CustomAssertions.assertClassHasMethod(
+                    generatedSource.getClassUnderTest(), "getValue"),
+                JsonValue.class);
+          }
         }
       }
     }
